fix: cleanup USE_FB2 macros.
1. Move uring source files under util/fibers.
2. Now third_party package builds with the same cxx compiler.
