+++ cmake/third_party.cmake
@@ -96,6 +96,7 @@ function(add_third_party name)
     CMAKE_ARGS -DCMAKE_ARCHIVE_OUTPUT_DIRECTORY:PATH=${INSTALL_ROOT}/lib
         -DCMAKE_LIBRARY_OUTPUT_DIRECTORY:PATH=${INSTALL_ROOT}/lib
         -DCMAKE_BUILD_TYPE:STRING=Release
+        -DCMAKE_CXX_COMPILER:STRING=${CMAKE_CXX_COMPILER}
         -DBUILD_TESTING=OFF
         "-DCMAKE_C_FLAGS:STRING=-O3" -DCMAKE_CXX_FLAGS=${THIRD_PARTY_CXX_FLAGS}
         -DCMAKE_INSTALL_PREFIX:PATH=${INSTALL_ROOT}
