+++ util/fibers/listener_interface.cc
@@ -7,14 +7,8 @@
 #include <signal.h>
 #include <sys/resource.h>
 
-// #include <boost/fiber/operations.hpp>
-
 #include "base/logging.h"
 #include "util/accept_server.h"
-#ifndef USE_FB2
-#include "util/fiber_sched_algo.h"
-#endif
-
 #include "util/proactor_pool.h"
 
 #define VSOCK(verbosity, sock) VLOG(verbosity) << "sock[" << native_handle(sock) << "] "
@@ -24,10 +18,6 @@ namespace util {
 
 using namespace boost;
 using namespace std;
-
-#ifndef USE_FB2
-using fibers_ext::Await;
-#endif
 
 namespace {
 
@@ -47,11 +37,7 @@ using ListType =
 
 struct ListenerInterface::TLConnList {
   ListType list;
-#ifdef USE_FB2
   fb2::CondVarAny empty_cv;
-#else
-  fibers::condition_variable_any empty_cv;
-#endif
 
   void Link(Connection* c) {
     DCHECK(!c->hook_.is_linked());
@@ -142,14 +128,6 @@ void ListenerInterface::RunAcceptLoop() {
   sock_->Shutdown(SHUT_RDWR);
   PreShutdown();
 
-#ifdef USE_FB2
-#else
-  // I am not a fan of "almost works" solutions. Here, I wait for Dispatch() calls
-  // to start running and decrease the probability we continue without the up to date conn_list.
-  // Since we are during the shutdown phase, waiting for 10ms is "probably" enough to let all the
-  // callback to unwind and run.
-  fibers_ext::SleepFor(10ms);
-#endif
   atomic_uint32_t cur_conn_cnt{0};
 
   pool_->AwaitFiberOnAll([&](auto* pb) {
